feat: modularize backend code and adjust express routing for lambda
